chore: added few changes

+++ UrlsGetByIdUpdateDelete.cs
@@ -1,9 +1,9 @@
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
 using Microsoft.Identity.Web.Resource;
 using Newtonsoft.Json;
 using UrlShortener.Function.Data;
@@ -16,10 +16,12 @@ namespace UrlShortener.Function
     public class UrlsGetByIdUpdateDelete
     {
         private readonly AppDbContext _context;
+        private readonly IHttpContextAccessor _contextAccessor;
 
-        public UrlsGetByIdUpdateDelete(AppDbContext context)
+        public UrlsGetByIdUpdateDelete(AppDbContext context, IHttpContextAccessor contextAccessor)
         {
             _context = context;
+            _contextAccessor = contextAccessor;
         }
 
         [Function("UrlsGetByIdUpdateDelete")]
@@ -28,6 +30,7 @@ namespace UrlShortener.Function
              "get","put","delete", Route = "urls/{id}")] HttpRequest req
             , int id)
         {
+            var user = _contextAccessor?.HttpContext?.User;
             if (req.Method == HttpMethods.Get)
             {
                 var url = await _context.Url.FirstOrDefaultAsync(u => u.Id == id);
@@ -51,6 +54,11 @@ namespace UrlShortener.Function
                     return new NotFoundResult();
                 }
 
+                if(existingUrl.UserId != user.FindFirst(ClaimTypes.NameIdentifier)?.Value)
+                {
+                    return new UnauthorizedResult();
+                }
+
                 existingUrl.Status = url.Status;
 
                 _context.Update(existingUrl);
@@ -65,6 +73,11 @@ namespace UrlShortener.Function
                 if (url == null)
                 {
                     return new NotFoundResult();
+                }
+
+                if(url.UserId != user.FindFirst(ClaimTypes.NameIdentifier)?.Value)
+                {
+                    return new UnauthorizedResult();
                 }
 
                 _context.Url.Remove(url);
